Refactor Git commands into GitHelper
